+++ shared/constants/config.js
@@ -1,9 +1,7 @@
 // @flow
 import uniq from 'lodash/uniq'
 import {isMobile, runMode} from './platform'
-import {type ConversationIDKey} from './types/chat'
-import {type Tab} from './tabs'
-import {type Config, type DeviceID, type ExtendedStatus} from './types/flow-types'
+import type {State} from './types/config'
 
 export const maxBootstrapTries = 3
 export const bootstrapRetryDelay = 10 * 1000
@@ -15,51 +13,11 @@ const defaultTeamPrefix = '/team/'
 // the installer.
 const readyForBootstrap = isMobile
 
-export type InitialState = {|
-  conversation?: ConversationIDKey,
-  tab?: Tab,
-  url?: string,
-|}
-
-export type BootStatus = 'bootStatusLoading' | 'bootStatusBootstrapped' | 'bootStatusFailure'
-
-export function privateFolderWithUsers(users: Array<string>): string {
-  return `${defaultKBFSPath}${defaultPrivatePrefix}${uniq(users).join(',')}`
-}
-
-export function publicFolderWithUsers(users: Array<string>): string {
-  return `${defaultKBFSPath}${defaultPublicPrefix}${uniq(users).join(',')}`
-}
-
-export function teamFolder(team: string): string {
-  return `${defaultKBFSPath}${defaultTeamPrefix}${team}`
-}
-
-// NOTE: All stores which go over the wire to remote windows CANNOT be immutable (yet)
-export type State = {
-  appFocused: boolean,
-  appFocusedCount: number,
-  bootStatus: BootStatus,
-  bootstrapTriesRemaining: number,
-  config: ?Config,
-  daemonError: ?Error,
-  deviceID: ?DeviceID,
-  deviceName: ?string,
-  error: ?any,
-  extendedConfig: ?ExtendedStatus,
-  followers: {[key: string]: true},
-  following: {[key: string]: true},
-  globalError: ?Error,
-  initialState: ?InitialState,
-  kbfsPath: string,
-  loggedIn: boolean,
-  pushLoaded: boolean,
-  readyForBootstrap: boolean,
-  registered: boolean,
-  uid: ?string,
-  userActive: boolean,
-  username: ?string,
-}
+export const privateFolderWithUsers = (users: Array<string>) =>
+  `${defaultKBFSPath}${defaultPrivatePrefix}${uniq(users).join(',')}`
+export const publicFolderWithUsers = (users: Array<string>) =>
+  `${defaultKBFSPath}${defaultPublicPrefix}${uniq(users).join(',')}`
+export const teamFolder = (team: string) => `${defaultKBFSPath}${defaultTeamPrefix}${team}`
 
 export const initialState: State = {
   appFocused: true,
